+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v0.8.1
+* Enhancements
+  * Adds leex and parsing for `and` and `or`.
+  * Adds leex and parsing for `!` and boolean.
+
 ## v0.8.0
   * New
     * `Predicator.matches?/3` accepts evaluator options.